+++ src/index.html
@@ -11,8 +11,6 @@
       rel="stylesheet"
     />
     <link rel="stylesheet" href="./sass/index.scss" />
-    <!-- <script type="module" src="./library.js"></script> -->
-    <!--Цей скріпт тут порібен?-->
     <script type="module" src="./index.js"></script>
   </head>
   <body>
